+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.0.3] - not released yet
 ### Added
+* support for [OpenGameArt](https://opengameart.org) assets
 * support for GIF & SVG images hosted on Wikimedia
 * extended range of supported Flickr URLs